Put dash app into separate file
